requirements.txt
PyPDF2
openai
python-dotenv
ipykernel
pymongo
fastapi
uvicorn
requests
python-multipart